refactored milestone note into component, fixed notes to update
